+++ examples/src/rescue_raps/custom_trace_table.rs
@@ -85,7 +85,7 @@ impl<B: StarkField> RapTraceTable<B> {
             "execution trace length must be a power of 2"
         );
         assert!(
-            log2(length) as u32 <= B::TWO_ADICITY,
+            log2(length) <= B::TWO_ADICITY,
             "execution trace length cannot exceed 2^{} steps, but was 2^{}",
             B::TWO_ADICITY,
             log2(length)
